Moved webapp/components/Surveys into webapp/components/survey/Surveys
