+++ webapp/components/survey/Surveys/HeaderLeft/HeaderLeft.js
