+++ webapp/views/App/views/Home/SurveyList/SurveyList.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Surveys } from '@webapp/components/Surveys'
+import { Surveys } from '@webapp/components/survey/Surveys'
 
 const SurveyList = () => (
   <Surveys module="surveys" moduleApiUri="/api/surveys" title="appModules.surveyList" showStatus />
