+++ webapp/views/App/views/Home/SurveyTemplateList/SurveyTemplateList.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Surveys } from '@webapp/components/Surveys'
+import { Surveys } from '@webapp/components/survey/Surveys'
 
 const SurveyTemplateList = () => (
   <Surveys module="surveyTemplates" moduleApiUri="/api/surveys" template title="appModules.surveyListTemplates" />
